LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/mutate-rules.https-expected.txt


FAIL Removing the <script type=speculationrules> must cancel the prefetch assert_equals: expected (undefined) undefined but got (string) "prefetch"
PASS Removing a rule from the <script type=speculationrules> must not cancel the prefetch
PASS Adding a rule to the <script type=speculationrules> must not prefetch
